Refactor linear module and remove xdg dependency
Refactor the `linear` module to consolidate functionality into `client.py`.
Remove the `xdg` dependency and related files. Update imports and references
accordingly.

Changes:
- Moved model definitions and functions from `models.py` and `queries.py` to `client.py`.
- Deleted `models.py`, `queries.py`, `gql.py`, and `xdg` related files.
- Updated `linear_cli.py` to import from `client.py`.
- Simplified `highlight.py` by removing the `wrap_text` function.
