--- src/linear/__init__.py
@@ -1,15 +1,0 @@
-from . import print
-from .models import ISSUE_STATES, Issue, Team, User, WorkflowState
-from .queries import get_issue, get_me, get_team
-
-__all__ = [
-    "User",
-    "WorkflowState",
-    "Issue",
-    "get_issue",
-    "get_me",
-    "print",
-    "get_team",
-    "Team",
-    "ISSUE_STATES",
-]
