--- src/xdg/__init__.py
@@ -1,17 +1,0 @@
-from .app_cache import XDGAppCache
-from .constants import (
-    XDG_CACHE_HOME,
-    XDG_CONFIG_HOME,
-    XDG_DATA_HOME,
-    XDG_RUNTIME_DIR,
-    XDG_STATE_HOME,
-)
-
-__all__ = [
-    "XDG_CACHE_HOME",
-    "XDG_CONFIG_HOME",
-    "XDG_DATA_HOME",
-    "XDG_RUNTIME_DIR",
-    "XDG_STATE_HOME",
-    "XDGAppCache",
-]
